+++ R/Lesson.R
@@ -36,6 +36,10 @@ Lesson <- R6::R6Class("Lesson",
     #'   files, default is `FALSE` for markdown files (deprecated and unused).
     rmd = FALSE,
 
+    #' @field overview \[`logical`\] when `TRUE`, the lesson is an overview
+    #'   lesson and does not necessarly contain any episodes. Defaults to `FALSE`
+    overview = FALSE,
+
     #' @description create a new Lesson object from a directory
     #' @param path \[`character`\] path to a lesson directory. This must have a
     #'   folder called `_episodes` within that contains markdown episodes. 
@@ -60,22 +64,13 @@ Lesson <- R6::R6Class("Lesson",
         jeky <- read_jekyll_episodes(path, rmd, ...)
         self$episodes <- jeky$episodes
         self$rmd <- jeky$rmd
+        self$overview <- jeky$overview
         self$sandpaper <- FALSE
       } else {
-        episode_path <- fs::path(path, "episodes")
-        extra_paths <- fs::path(path, c("instructors", "learners", "profiles"))
-        cfg <- fs::dir_ls(path, regexp = "config[.]ya?ml")
-
-        self$episodes <- read_markdown_files(
-          episode_path, cfg, process_tags = FALSE, ...)
-
-        standard_files <- read_markdown_files(path, process_tags = FALSE, ...)
-
-        extra_files <- purrr::flatten(purrr::map(extra_paths,
-          read_markdown_files, cfg, process_tags = FALSE, ...))
-
-        self$extra <- c(standard_files, extra_files)
-
+        sandy <- read_sandpaper_lesson(path, ...)
+        self$episodes <- sandy$episodes
+        self$extra <- sandy$extra
+        self$overview <- sandy$overview
       }
       self$path <- path
     },
@@ -304,8 +299,11 @@ Lesson <- R6::R6Class("Lesson",
     #' frg <- Lesson$new(lesson_fragment())
     #' frg$validate_headings()
     validate_headings = function(verbose = TRUE) {
-      res <- purrr::map(self$episodes, 
-        ~.x$validate_headings(verbose = verbose, warn = FALSE)
+      res <- purrr::map(c(self$episodes, self$extra), 
+        function(x) {
+          if (startsWith(x$name, "README")) return(NULL)
+          x$validate_headings(verbose = verbose, warn = FALSE)
+        }
       )
       res <- stack_rows(res)
       throw_heading_warnings(res)
@@ -330,7 +328,12 @@ Lesson <- R6::R6Class("Lesson",
     #' frg <- Lesson$new(lesson_fragment())
     #' frg$validate_divs()
     validate_divs = function() {
-      res <- purrr::map(self$episodes, ~.x$validate_divs(warn = FALSE))
+      res <- purrr::map(c(self$episodes, self$extra), 
+        function(x) {
+          if (startsWith(x$name, "README")) return(NULL)
+          x$validate_divs(warn = FALSE)
+        }
+      )
       res <- stack_rows(res)
       throw_div_warnings(res)
       invisible(res)
@@ -359,7 +362,12 @@ Lesson <- R6::R6Class("Lesson",
     #' frg <- Lesson$new(lesson_fragment())
     #' frg$validate_links()
     validate_links = function() {
-      res <- purrr::map(self$episodes, ~.x$validate_links(warn = FALSE))
+      res <- purrr::map(c(self$episodes, self$extra),
+        function(x) {
+          if (startsWith(x$name, "README")) return(NULL)
+          x$validate_links(warn = FALSE)
+        }
+      )
       res <- stack_rows(res)
       throw_link_warnings(res)
       invisible(res)