+++ tests/testthat/test-Lesson.R
@@ -22,6 +22,109 @@ test_that("Sandpaper lessons can be read", {
   expect_s3_class(snd$episodes[[1]], "Episode")
   expect_null(snd$built)
   expect_s3_class(snd$extra[[1]], "Episode")
+  expect_false(snd$overview)
+})
+
+test_that("Sandpaper lessons with incomplete conversions will throw an error", {
+  tmp <- withr::local_tempdir()
+
+  test_dir <- fs::path(tmp, "ohno")
+  fs::dir_copy(lesson_fragment("sandpaper-fragment"), test_dir)
+  fs::dir_copy(fs::path(frg_path, "_episodes"), test_dir)
+  fs::file_copy(fs::path(frg_path, "_config.yml"), test_dir)
+    
+  expect_error(Lesson$new(path = test_dir, jekyll = FALSE), 
+    "config files in the lesson.+?_config.yml")
+})
+
+
+test_that("Jekyll workshop overview lessons with no episodes can be read", {
+  tmp <- withr::local_tempdir()
+
+  # the key is that they end with -workshop
+  test_dir <- fs::path(tmp, "jekyll-test-workshop")
+  fs::dir_copy(frg_path, test_dir)
+  fs::dir_delete(fs::path(test_dir, "_episodes"))
+  # We are expecting this to _not fail_
+  expect_failure(expect_error({
+    lsn <- Lesson$new(path = test_dir, jekyll = TRUE)
+  }))
+  expect_true(lsn$overview)
+  # the lesson should have an empty episode slot
+  expect_null(lsn$episodes)
+  lnks <- lsn$validate_links()
+
+  expect_s3_class(lnks, "data.frame")
+  expect_equal(nrow(lnks), 0L)
+})
+
+test_that("Jekyll workshop overview lessons with episodes can be read (but won't be overview)", {
+  tmp <- withr::local_tempdir()
+
+  # the key is that they end with -workshop
+  test_dir <- fs::path(tmp, "jekyll-test-workshop")
+  fs::dir_copy(frg_path, test_dir)
+  # We are expecting this to _not fail_
+  expect_failure(expect_error({
+    lsn <- Lesson$new(path = test_dir, jekyll = TRUE)
+  }))
+
+  expect_false(lsn$overview)
+  # the episodes should still exist for an overview lesson
+  expect_type(lsn$episodes, "list")
+  expect_length(lsn$episodes, 4L)
+  expect_s3_class(lsn$episodes[[1]], "Episode")
+
+  # the lesson should throw warnings about the missing files in the lesson
+  # fragment demo because the episodes will still exist.
+  suppressMessages(expect_message(lnks <- lsn$validate_links(), "missing file"))
+
+  expect_s3_class(lnks, "data.frame")
+  expect_equal(nrow(lnks), 14L)
+})
+
+test_that("Sandpaper workshop overview lessons with no episodes can be read", {
+  tmp <- withr::local_tempdir()
+
+  # the key is that they end with -workshop
+  test_dir <- fs::path(tmp, "sandpaper-test-workshop")
+  fs::dir_copy(lesson_fragment("sandpaper-fragment"), test_dir)
+  fs::dir_delete(fs::path(test_dir, "episodes"))
+  cat("\noverview: true\n", 
+    file = fs::path(test_dir, "config.yaml"), append = TRUE)
+  # We are expecting this to _not fail_
+  expect_failure(expect_error({
+    lsn <- Lesson$new(path = test_dir, jekyll = FALSE)
+  }))
+  # the lesson should have an empty episode slot
+  expect_null(lsn$episodes)
+  expect_true(lsn$overview)
+
+  # the setup page should throw warnings about two HTTP links
+  suppressMessages(expect_message(lnks <- lsn$validate_links(), "HTTPS"))
+
+  expect_s3_class(lnks, "data.frame")
+  expect_equal(nrow(lnks), 2L)
+})
+
+
+if (requireNamespace("cli")) {
+  cli::test_that_cli("Sandpaper Lessons can be validated", {
+    snd <- Lesson$new(path = lesson_fragment("sandpaper-fragment"), jekyll = FALSE)
+    withr::local_envvar(list(CI = 'true'))
+    expect_snapshot(vhead <- snd$validate_headings())
+    expect_equal(nrow(vhead), 8L)
+    expect_snapshot(vlink <- snd$validate_links())
+    expect_equal(nrow(vlink), 3L)
+  })
+}
+
+test_that("Sandpaper Lessons can be _quietly_ validated", {
+  snd <- Lesson$new(path = lesson_fragment("sandpaper-fragment"), jekyll = FALSE)
+  suppressMessages({
+  expect_message(vhead <- snd$validate_headings(verbose = FALSE), NA)
+  expect_message(vlink <- snd$validate_links(), "There were errors in 2/3 links")
+  })
 })
 
 test_that("Sandpaper lessons have getter and summary methods", {
@@ -103,6 +206,7 @@ test_that("Lesson class contains the right stuff", {
   expect_equal(frg$path, unique(purrr::map_chr(frg$episodes, "lesson")))
   expect_equal(frg$n_problems, c("10-lunch.md" = 0, "12-for-loops.md" = 0, "14-looping-data-sets.md" = 0, "17-scope.md" = 0))
   expect_length(frg$show_problems, 0)
+  expect_false(frg$overview)
 
 })
 
@@ -112,14 +216,14 @@ if (requireNamespace("cli")) {
     expect_snapshot(vhead <- frg$validate_headings())
     expect_equal(nrow(vhead), 37L)
     expect_snapshot(vlink <- frg$validate_links())
-    expect_equal(nrow(vlink), 13L)
+    expect_equal(nrow(vlink), 14L)
   })
 }
 
 test_that("Lessons can be _quietly_ validated", {
   suppressMessages({
   expect_message(vhead <- frg$validate_headings(verbose = FALSE), "There were errors in 13/37 headings")
-  expect_message(vlink <- frg$validate_links(), "There were errors in 3/13 links")
+  expect_message(vlink <- frg$validate_links(), "There were errors in 4/14 links")
   })
 })